Split the demo composition into scene modules
video/src/VibeadsDemo.tsx held the whole 30-second Remotion demo in one
1263-line file: a single 864-line component built from eight scene-sized
IIFEs. It broke four oxlint agent budgets (max-lines, two counts of
max-lines-per-function, and complexity on TypingText).

Split it along the seams the code already had. Colours move to theme.ts
and frame numbers to timings.ts. The six presentational components get
their own files, and each of the eight scenes becomes a component, with
scenes 5 and 6 further split into their four placements and four stat
blocks. VibeadsDemo is now a scene table plus the two audio tracks.

Every scene takes `frame` as a prop rather than calling useCurrentFrame().
That is load-bearing: the hook is called once, outside every <Sequence>,
so the scenes animate off the absolute frame. A scene that called the
hook itself would sit inside a <Sequence> and get a sequence-relative
frame, silently shifting every animation.

Two smaller changes fall out of the split: TypingText hands its blinking
cursor to a TypingCursor child, which drops its complexity from 11 to
under the cap, and the unused Easing import and unused fps binding go.

Verified by rendering all 900 frames to PNG before and after and
comparing sha256 per frame: 900/900 identical, 694 of them distinct. The
harness itself renders twice from a clean bundle and is deterministic.

With the file clean, promote complexity, max-lines and
max-lines-per-function from warn to error so regressions fail CI.
